backend/requirements.txt
ultralytics
Pillow
requests